Use dogpile.cache to vastly speed up the app.
Currently, every page load in the voting process requires that the
elections make ``N`` queries to FAS where ``N`` is the number of
candidates.  It's being really slow in production tonight with the
opening of the Council election.

This should add some generic caching so that we can use an on-disk .dbm
cache if we like or alternatively memcached.

Tested locally.